+++ es-core/src/Window.h
@@ -61,7 +61,7 @@ public:
 	void pushGui(GuiComponent* gui);
 	void removeGui(GuiComponent* gui);
 	GuiComponent* peekGui();
-	inline int getGuiStackSize() { return mGuiStack.size(); }
+	inline int getGuiStackSize() { return (int)mGuiStack.size(); }
 
 	void textInput(const char* text);
 	void input(InputConfig* config, Input input);
