+++ es-core/src/math/Vector3f.cpp
@@ -2,9 +2,9 @@
 
 Vector3f& Vector3f::round()
 {
-	mX = (int)(mX + 0.5f);
-	mY = (int)(mY + 0.5f);
-	mZ = (int)(mZ + 0.5f);
+	mX = (float)(int)(mX + 0.5f);
+	mY = (float)(int)(mY + 0.5f);
+	mZ = (float)(int)(mZ + 0.5f);
 	
 	return *this;
