+++ es-core/src/math/Vector4f.cpp
@@ -2,10 +2,10 @@
 
 Vector4f& Vector4f::round()
 {
-	mX = (int)(mX + 0.5f);
-	mY = (int)(mY + 0.5f);
-	mZ = (int)(mZ + 0.5f);
-	mW = (int)(mW + 0.5f);
+	mX = (float)(int)(mX + 0.5f);
+	mY = (float)(int)(mY + 0.5f);
+	mZ = (float)(int)(mZ + 0.5f);
+	mW = (float)(int)(mW + 0.5f);
 	
 	return *this;
